Beta release with working state-tray and bugfixes for config file corruptions

+++ pkg/platform/runtime/setup/implementations/camel/runtime.go
@@ -18,12 +18,20 @@ func NewSetup(s *store.Store) *Setup {
 	return &Setup{s}
 }
 
-// ReusableArtifacts returns an empty, because camel installations cannot re-use and artifacts from previous installations
-func (s *Setup) ReusableArtifacts(_ artifact.ArtifactChangeset, _ store.StoredArtifactMap) store.StoredArtifactMap {
+// ReusableArtifacts returns an empty map, because camel installations cannot re-use and artifacts from previous installations
+// Only when the recipe is identical, it returns the stored artifact map.  This can happen when the executors director needs to be updated.
+func (s *Setup) ReusableArtifacts(changed artifact.ArtifactChangeset, stored store.StoredArtifactMap) store.StoredArtifactMap {
+	if len(changed.Added) == 0 && len(changed.Updated) == 0 && len(changed.Removed) == 0 {
+		return stored
+	}
 	return make(store.StoredArtifactMap)
 }
 
-func (s *Setup) DeleteOutdatedArtifacts(_ artifact.ArtifactChangeset, _, _ store.StoredArtifactMap) error {
+// DeleteOutdatedArtifacts deletes the entire installation directory, unless alreadyInstalled is not zero, which can happen when the executors directory needs to be re-generated.
+func (s *Setup) DeleteOutdatedArtifacts(_ artifact.ArtifactChangeset, _, alreadyInstalled store.StoredArtifactMap) error {
+	if len(alreadyInstalled) != 0 {
+		return nil
+	}
 	if err := os.RemoveAll(s.store.InstallPath()); err != nil {
 		logging.Error("Error removing previous camel installation: %v", err)
 	}
